Clean up unit config loading
Make the unit-projected global configuration the normal load path and
split config models, projection, and validation into focused files so
phase 7 additions have clearer boundaries.
